move object utilities
